src/requirements.txt
# Chess Commentary Training Pipeline - Requirements
# Install with: pip install -r requirements.txt

# Core ML
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
torch 
torchvision --index-url https://download.pytorch.org/whl/cu126
wandb

# ONNX Runtime
onnx>=1.15.0
#onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  # Uncomment for GPU support

# Chess
python-chess>=1.10.0

# Data processing/generation
google-genai #if using gemini
PyQt6

maia2>=0.1.0
gdown
pyzstd
pyyaml
pandas
einops


tqdm>=4.65.0
requests>=2.31.0


# Optional: for visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
